✨ chore: Upgrade testing packages to latest versions
Update the project file `ACDClinicManagement.Tests.csproj` to
reference newer versions of MSTest and Microsoft Testing Extensions.
Upgrade from version 1.7.2 to 1.7.3 for several Microsoft Testing
packages and from version 3.9.2 to 3.9.3 for MSTest packages. This
ensures compatibility with the latest features and fixes provided
by these libraries.

Also update the `packages.config` file to reflect the new versions
of the packages, ensuring that the project uses the latest
dependencies for testing.

Saeid Ghasemi