🚧 fix: update application URLs and correct product name
Update the `PublishUrl` and `InstallUrl` to reflect the new
directory for the application, changing from
`acdclinicmanagement` to `ophthalmology_clinic_management`.

Correct the `ProductName` from "Ophthalmology Office Management"
to "Ophthalmology Clinic Management Syestem", addressing a
spelling error in the product name.

These changes ensure that the application points to the correct
resources and presents the correct product name to users.

Saeid Ghasemi